Fixed navigation menu bug.
When the navigation menu is open, the html DOM element is given a
style of "overflow: hidden;". When closed via the ToggleMenu action,
the style is removed.

However, the TouchEnd action also had the ability to close the menu,
but it did not removed the overflow style.

To fix this, I'm using UpdateWithSideEffects and sending a ToggleMenu
action instead of updating the nav state directly.