Refactor
Refactored the `route` type to be a record. This results in nicer user
code in App.re and removes the need for Config.re.

Also replaced App state to use a record to store touches instead of a
list. There's no need to store every touch event's position; only the
first and last touch events.